refactor: move SudokuListItem and AlarmReceiver out of ui package
SudokuListItem was a plain Kotlin presentation model with no Android/UI
dependency, so it belongs in domain.model. AlarmReceiver is a
BroadcastReceiver (an Android entry point, not an Activity/Fragment/View/
ViewModel), so it gets its own receivers package sibling to ui/domain/data.

This removes the last domain-layer imports of ui.*, so the
ArchitectureTest exceptions for ObserveDailySudokusUseCase,
ObserveSudokuHistoryUseCase, ObserveSudokuLevelUseCase, and
SendDailyNotificationUseCase are no longer needed.
