+++ app/src/main/java/de/lemke/sudoku/domain/ObserveSudokuLevelUseCase.kt
@@ -17,7 +17,7 @@
 package de.lemke.sudoku.domain
 
 import de.lemke.sudoku.data.database.SudokusRepository
-import de.lemke.sudoku.ui.utils.SudokuListItem
+import de.lemke.sudoku.domain.model.SudokuListItem
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
